Move PixmapScaling to kas::image::Scaling
