+++ crates/kas-image/src/lib.rs
@@ -25,6 +25,12 @@ pub use canvas::{Canvas, CanvasProgram};
 pub use image::{Image, ImageError};
 #[cfg(feature = "svg")] pub use svg::Svg;
 
+use kas::cast::{Conv, ConvFloat};
+use kas::geom::{Rect, Vec2};
+use kas::impl_scope;
+use kas::layout::{AlignPair, AxisInfo, LogicalSize, SizeRules, Stretch};
+use kas::theme::{MarginStyle, SizeCx};
+
 /// Load a window icon from a path
 #[cfg(feature = "image")]
 pub fn window_icon_from_path<P: AsRef<std::path::Path>>(
@@ -38,4 +44,79 @@ pub fn window_icon_from_path<P: AsRef<std::path::Path>>(
         .into_rgba8();
     let (w, h) = im.dimensions();
     Ok(kas::window::Icon::from_rgba(im.into_vec(), w, h)?)
+}
+
+impl_scope! {
+    /// Control over image scaling
+    #[impl_default]
+    #[derive(Clone, Debug, PartialEq)]
+    pub struct Scaling {
+        /// Margins
+        pub margins: MarginStyle,
+        /// Display size
+        ///
+        /// This may be set by the providing type or by the user.
+        pub size: LogicalSize,
+        /// Minimum size relative to [`Self::size`]
+        ///
+        /// Default: `1.0`
+        pub min_factor: f32 = 1.0,
+        /// Ideal size relative to [`Self::size`]
+        ///
+        /// Default: `1.0`
+        pub ideal_factor: f32 = 1.0,
+        /// If true, aspect ratio is fixed relative to [`Self::size`]
+        ///
+        /// Default: `true`
+        pub fix_aspect: bool = true,
+        /// Widget stretchiness
+        ///
+        /// If is `None`, max size is limited to ideal size.
+        pub stretch: Stretch,
+        /// The assigned [`Rect`]
+        pub rect: Rect,
+    }
+}
+
+impl Scaling {
+    /// Generates [`SizeRules`] based on size
+    pub fn size_rules(&mut self, sizer: SizeCx, axis: AxisInfo) -> SizeRules {
+        let margins = sizer.margins(self.margins).extract(axis);
+        let scale_factor = sizer.scale_factor();
+        let min = self
+            .size
+            .to_physical(scale_factor * self.min_factor)
+            .extract(axis);
+        let ideal = self
+            .size
+            .to_physical(scale_factor * self.ideal_factor)
+            .extract(axis);
+        SizeRules::new(min, ideal, margins, self.stretch)
+    }
+
+    /// Constrains and aligns within `rect`
+    ///
+    /// The resulting size is then aligned using the `align` hints, defaulting to centered.
+    pub fn set_rect(&mut self, rect: Rect, align: AlignPair, scale_factor: f32) {
+        let mut size = rect.size;
+
+        if self.stretch == Stretch::None {
+            let ideal = self.size.to_physical(scale_factor * self.ideal_factor);
+            size = size.min(ideal);
+        }
+
+        if self.fix_aspect {
+            let logical_size = Vec2::from(self.size);
+            let Vec2(rw, rh) = Vec2::conv(size) / logical_size;
+
+            // Use smaller ratio, if any is finite
+            if rw < rh {
+                size.1 = i32::conv_nearest(rw * logical_size.1);
+            } else if rh < rw {
+                size.0 = i32::conv_nearest(rh * logical_size.0);
+            }
+        }
+
+        self.rect = align.aligned_rect(size, rect);
+    }
 }
